removed duplicate test
