+++ metron-platform/metron-elasticsearch/src/test/java/org/apache/metron/elasticsearch/dao/ElasticsearchMetaAlertDaoTest.java
@@ -18,32 +18,12 @@
 
 package org.apache.metron.elasticsearch.dao;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
-
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-import org.apache.metron.common.Constants;
 import org.apache.metron.indexing.dao.AccessConfig;
 import org.apache.metron.indexing.dao.HBaseDao;
 import org.apache.metron.indexing.dao.IndexDao;
 import org.apache.metron.indexing.dao.MultiIndexDao;
 import org.apache.metron.indexing.dao.metaalert.MetaAlertConfig;
-import org.apache.metron.indexing.dao.metaalert.MetaAlertConstants;
 import org.apache.metron.indexing.dao.metaalert.MetaAlertCreateRequest;
-import org.apache.metron.indexing.dao.metaalert.MetaAlertDao;
-import org.apache.metron.indexing.dao.metaalert.MetaScores;
 import org.apache.metron.indexing.dao.search.FieldType;
 import org.apache.metron.indexing.dao.search.GetRequest;
 import org.apache.metron.indexing.dao.search.GroupRequest;
@@ -55,6 +35,18 @@ import org.apache.metron.indexing.dao.update.CommentAddRemoveRequest;
 import org.apache.metron.indexing.dao.update.Document;
 import org.elasticsearch.index.IndexNotFoundException;
 import org.junit.Test;
+
+import java.io.IOException;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.verify;
 
 public class ElasticsearchMetaAlertDaoTest {
 
@@ -147,71 +139,6 @@ public class ElasticsearchMetaAlertDaoTest {
     MetaAlertCreateRequest createRequest = new MetaAlertCreateRequest();
     createRequest.setAlerts(Collections.singletonList(new GetRequest("don't", "care")));
     emaDao.createMetaAlert(createRequest);
-  }
-
-  @Test
-  public void testCalculateMetaScoresList() {
-    final double delta = 0.001;
-    List<Map<String, Object>> alertList = new ArrayList<>();
-
-    // add an alert with a threat score
-    alertList.add( Collections.singletonMap(ElasticsearchMetaAlertDao.THREAT_TRIAGE_FIELD, 10.0f));
-
-    // add a second alert with a threat score
-    alertList.add( Collections.singletonMap(ElasticsearchMetaAlertDao.THREAT_TRIAGE_FIELD, 20.0f));
-
-    // add a third alert with NO threat score
-    alertList.add( Collections.singletonMap("alert3", "has no threat score"));
-
-    // create the metaalert
-    Map<String, Object> docMap = new HashMap<>();
-    docMap.put(MetaAlertConstants.ALERT_FIELD, alertList);
-    Document metaalert = new Document(docMap, "guid", MetaAlertConstants.METAALERT_TYPE, 0L);
-
-    // calculate the threat score for the metaalert
-    MetaScores.calculateMetaScores(metaalert, MetaAlertConstants.THREAT_FIELD_DEFAULT, MetaAlertConstants.THREAT_SORT_DEFAULT);
-    Object threatScore = metaalert.getDocument().get(ElasticsearchMetaAlertDao.THREAT_TRIAGE_FIELD);
-
-    // the metaalert must contain a summary of all child threat scores
-    assertEquals(20D, (Double) metaalert.getDocument().get("max"), delta);
-    assertEquals(10D, (Double) metaalert.getDocument().get("min"), delta);
-    assertEquals(15D, (Double) metaalert.getDocument().get("average"), delta);
-    assertEquals(2L, metaalert.getDocument().get("count"));
-    assertEquals(30D, (Double) metaalert.getDocument().get("sum"), delta);
-    assertEquals(15D, (Double) metaalert.getDocument().get("median"), delta);
-
-    // it must contain an overall threat score; a float to match the type of the threat score of the other sensor indices
-    assertTrue(threatScore instanceof Float);
-
-    // by default, the overall threat score is the sum of all child threat scores
-    assertEquals(30.0F, threatScore);
-  }
-
-  @Test
-  public void testCalculateMetaScoresWithDifferentFieldName() {
-    List<Map<String, Object>> alertList = new ArrayList<>();
-
-    // add an alert with a threat score
-    alertList.add( Collections.singletonMap(MetaAlertConstants.THREAT_FIELD_DEFAULT, 10.0f));
-
-    // create the metaalert
-    Map<String, Object> docMap = new HashMap<>();
-    docMap.put(MetaAlertConstants.ALERT_FIELD, alertList);
-    Document metaalert = new Document(docMap, "guid", MetaAlertConstants.METAALERT_TYPE, 0L);
-
-    // Configure a different threat triage score field name
-    AccessConfig accessConfig = new AccessConfig();
-    accessConfig.setGlobalConfigSupplier(() -> new HashMap<String, Object>() {{
-      put(Constants.THREAT_SCORE_FIELD_PROPERTY, MetaAlertConstants.THREAT_FIELD_DEFAULT);
-    }});
-    ElasticsearchDao elasticsearchDao = new ElasticsearchDao();
-    elasticsearchDao.setAccessConfig(accessConfig);
-
-    // calculate the threat score for the metaalert
-    ElasticsearchMetaAlertDao metaAlertDao = new ElasticsearchMetaAlertDao();
-    metaAlertDao.init(elasticsearchDao);
-    MetaScores.calculateMetaScores(metaalert, MetaAlertConstants.THREAT_FIELD_DEFAULT, MetaAlertConstants.THREAT_SORT_DEFAULT);
-    assertNotNull(metaalert.getDocument().get(MetaAlertConstants.THREAT_FIELD_DEFAULT));
   }
 
   @Test
